+++ test/triangle/main.cpp
@@ -599,7 +599,7 @@ struct App {
         atts[1].format = depthFormat;
         atts[1].samples = samples;
         atts[1].loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-        atts[1].storeOp = VK_ATTACHMENT_STORE_OP_STORE;
+        atts[1].storeOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;   // nothing reads the depth buffer after the pass
         atts[1].stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
         atts[1].stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
         atts[1].initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;